Commit files level up
