--- Uitnodigingsregel/data/interim/train_data_standardized.txt
@@ -1,1 +1,0 @@
-
