--- Uitnodigingsregel/data/processed/ranked_students_rf.csv
@@ -1,151 +1,0 @@
-ranking	Studentnummer	voorspelling
-1.0	418236	0.1861983798307181
-2.0	423953	0.17888934659533662
-3.0	424854	0.17840217892146285
-4.0	428138	0.17703994315623667
-5.0	420088	0.17700701472550798
-6.0	419658	0.17699050267976976
-7.0	424182	0.17599377930183713
-8.0	421528	0.17431997243572103
-9.0	425787	0.17388394304823002
-10.0	420728	0.17101741166124312
-11.0	421065	0.17087845151013403
-12.0	424088	0.1705840498301744
-13.0	417649	0.16967014176986703
-14.0	419573	0.16921056552639435
-15.0	424933	0.1685958785810484
-16.0	427194	0.16856322460164205
-17.0	423957	0.16840305872622427
-18.0	417520	0.16839649402677398
-19.0	424693	0.16772691408500826
-20.0	416450	0.16766915668085136
-21.0	420255	0.16764179334835166
-22.0	425755	0.16761065991725296
-23.0	419013	0.16726878930004876
-24.0	419374	0.16718895613122853
-25.0	426386	0.16647716331963003
-26.0	421860	0.16616651242487893
-27.0	422763	0.16582221067498723
-28.0	421569	0.16576633026477883
-29.0	419841	0.16562268958732818
-30.0	419202	0.16549972608851413
-31.0	421222	0.16500674383466196
-32.0	423109	0.16499297754775125
-33.0	424600	0.16478130310923012
-34.0	422721	0.1646515515827123
-35.0	420847	0.16431396507085555
-36.0	426752	0.16421578989341148
-37.0	427839	0.1638042812637529
-38.0	422290	0.16374750744082933
-39.0	421968	0.16361945600207622
-40.0	386411	0.1634783689622514
-41.0	421426	0.1633401603315727
-42.0	419035	0.1632280969650611
-43.0	420623	0.163213647408314
-44.0	415458	0.16299596101451827
-45.0	425287	0.1628314157233703
-46.0	423151	0.1626403142193264
-47.0	428538	0.16252141495420364
-48.0	417133	0.16248643597452542
-49.0	427326	0.16241685182757426
-50.0	418469	0.1622184360531801
-51.0	424333	0.16216556712249575
-52.0	418404	0.16215845831555156
-53.0	429003	0.16213780923958507
-54.0	427058	0.16184298293305
-55.0	423053	0.16173875704740012
-56.0	423622	0.1615787426870063
-57.0	422360	0.16145662147593431
-58.0	420450	0.1613532841869907
-59.0	419833	0.16127468788654425
-60.0	421286	0.16122387793995027
-61.0	424968	0.1608619279600657
-62.0	425961	0.1606191828440005
-63.0	426453	0.16061524824780787
-64.0	422113	0.160505985056618
-65.0	420468	0.16043544980910668
-66.0	425998	0.16027518052389586
-67.0	424201	0.16009939876262608
-68.0	421440	0.15990449643854393
-69.0	420371	0.15980199900363937
-70.0	426278	0.15973479816375494
-71.0	427646	0.15967659124427727
-72.0	418048	0.15963171678229485
-73.0	425621	0.15957944256222922
-74.0	421762	0.15952999872262924
-75.0	423873	0.1593925681282465
-76.0	418050	0.15926293603571204
-77.0	418780	0.15912172347103215
-78.0	425447	0.15911684657772798
-79.0	423912	0.15901490905541002
-80.0	423250	0.15890842133651503
-81.0	419836	0.15872022417735296
-82.0	422126	0.15863988216121513
-83.0	423132	0.1585760504307596
-84.0	419754	0.15857418006715462
-85.0	420215	0.15856256912140076
-86.0	422840	0.15850949886392818
-87.0	422950	0.1584360076439211
-88.0	431227	0.1583147566038552
-89.0	419315	0.1580412833356487
-90.0	420776	0.157802844812141
-91.0	419186	0.1577948294033168
-92.0	421808	0.15763736476976742
-93.0	424557	0.15749576296878517
-94.0	419893	0.15744517871025476
-95.0	392049	0.1572397761501716
-96.0	422029	0.15684446479276848
-97.0	423241	0.1564603860452402
-98.0	420701	0.15614631066446116
-99.0	422953	0.15607201977093924
-100.0	420455	0.15599572376023205
-101.0	421377	0.15581615968132112
-102.0	422603	0.15548791110905194
-103.0	426342	0.15542292446573303
-104.0	418852	0.15512799164519606
-105.0	423494	0.15477810209790788
-106.0	424795	0.1547113840938223
-107.0	425587	0.15467484367455556
-108.0	422331	0.15461898967630008
-109.0	422881	0.15456207821986045
-110.0	417660	0.15453864029522504
-111.0	421214	0.15405834432574975
-112.0	419477	0.15392197047081718
-113.0	424453	0.15383187594821995
-114.0	417593	0.15353827977252377
-115.0	427927	0.1534202166128315
-116.0	418458	0.1533378164070605
-117.0	419383	0.15332165283217608
-118.0	420230	0.15324650811936225
-119.0	427990	0.1531938686048582
-120.0	429002	0.15315368208489738
-121.0	348081	0.1530826219515535
-122.0	417488	0.15303922263857403
-123.0	422265	0.15281798511559053
-124.0	371312	0.1526737121534495
-125.0	418773	0.15264315758582758
-126.0	419669	0.1524919427645304
-127.0	421332	0.15248416209438118
-128.0	428439	0.15230180599726245
-129.0	426618	0.15227161189659055
-130.0	425400	0.15205317700869764
-131.0	421491	0.1514688586412516
-132.0	424849	0.15129876512311127
-133.0	425481	0.15125296645999678
-134.0	420684	0.15108525753762228
-135.0	423760	0.15093192181668758
-136.0	426852	0.15083262513715398
-137.0	422084	0.15044402359294426
-138.0	420611	0.1504051426379685
-139.0	423217	0.15038035031525163
-140.0	423072	0.14994149513764501
-141.0	425559	0.14947139211202304
-142.0	424940	0.14946557549930825
-143.0	420655	0.14937569788079874
-144.0	418740	0.14872501852585782
-145.0	420968	0.1480384664908491
-146.0	424883	0.14741231827564943
-147.0	408316	0.1463243336039896
-148.0	431255	0.1440955292782964
-149.0	422124	0.14377421362206175
-150.0	430925	0.13819611788801134
