--- Uitnodigingsregel/docs/docs/index.md
@@ -1,10 +1,0 @@
-# Uitnodigingsregel documentation!
-
-## Description
-
-Studenten uitval voorspellen
-
-## Commands
-
-The Makefile contains the central entry points for common tasks related to this project.
-
