--- Uitnodigingsregel/docs/mkdocs.yml
@@ -1,4 +1,0 @@
-site_name: Uitnodigingsregel
-# 
-site_author: Mondriaan
-#
