--- Uitnodigingsregel/environment.yml
@@ -1,23 +1,0 @@
-name: Uitnodigingsregel
-channels:
-  - conda-forge
-dependencies:
-  - python=3.13
-  - black
-  - flake8
-  - isort
-  - pip
-  - typer
-  - loguru
-  - tqdm
-  - ipython
-  - jupyterlab
-  - matplotlib
-  - notebook
-  - numpy
-  - pandas
-  - scikit-learn
-  - pip:
-    - python-dotenv
-    - mkdocs
-    - -e .
