--- Uitnodigingsregel/pyproject.toml
@@ -1,32 +1,0 @@
-[build-system]
-requires = ["flit_core >=3.2,<4"]
-build-backend = "flit_core.buildapi"
-
-[project]
-name = "Module"
-version = "0.0.1"
-description = "Studenten uitval voorspellen"
-authors = [
-  { name = "Mondriaan" },
-]
-license = { file = "LICENSE" }
-readme = "README.md"
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: MIT License"
-]
-requires-python = "~=3.13"
-
-[tool.black]
-line-length = 99
-include = '\.pyi?$'
-exclude = '''
-/(
-    \.git
-  | \.venv
-)/
-'''
-
-[tool.ruff.lint.isort]
-known_first_party = ["Module"]
-force_sort_within_sections = true
