+++ docs/README.md
@@ -1,0 +1,12 @@
+Generating the docs
+----------
+
+Use [mkdocs](http://www.mkdocs.org/) structure to update the documentation. 
+
+Build locally with:
+
+    mkdocs build
+
+Serve locally with:
+
+    mkdocs serve
