+++ docs/docs/getting-started.md
@@ -1,0 +1,6 @@
+Getting started
+===============
+
+This is where you describe how to get set up on a clean install, including the
+commands necessary to get the raw data (using the `sync_data_from_s3` command,
+for example), and then how to make the cleaned, final data sets.
