+++ docs/mkdocs.yml
@@ -1,0 +1,4 @@
+site_name: Uitnodigingsregel
+# 
+site_author: Mondriaan
+#
