+++ setup.cfg
@@ -1,0 +1,4 @@
+[flake8]
+ignore = E731,E266,E501,C901,W503
+max-line-length = 99
+exclude = .git,notebooks,references,models,data
